add format config files
format changed source files
extend test script to check snmp values too

+++ simple-snmp-test.sh
@@ -1,21 +1,37 @@
 #!/bin/bash
+
+dir_name=`basename "$0"`
+export MIBDIRS=+${dir_name}/mibs
+export MIBS=ALL
+
 set -u
 set -e
 set -x
 
-#TBD rm -rf config
-#TBD rm -f snmpv3_boot_counter
+# cleanup config files
+rm -f snmpv3_boot_counter
+rm -rf config
 
 mkdir -p config
 
 killall agent || echo OK
 pkill agent || echo OK
 
-examples/static_table/src/agent &
+# start as bg job
+examples/static_table/src/agent 4700 &
 sleep 1
 cat snmpv3_boot_counter
 
 snmpstatus -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4700
+
+snmpEngineBoots=`snmpget -v1 -c public -Onqv localhost:4700 snmpEngineBoots.0`
+test ${snmpEngineBoots} -eq 1 || exit 1
+echo "OK, first boot"
+
+snmpOutTraps=`snmpget -v1 -c public -Onqv localhost:4700 snmpOutTraps.0`
+test ${snmpOutTraps} -eq 1 || exit 1
+echo "OK, cold start trap sent"
+
 snmpset -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4700 snmpEnableAuthenTraps.0 = enabled
 
 snmpwalk -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4700 snmp
@@ -24,8 +40,8 @@ snmpwalk -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4700 snmpEngine
 
 snmpbulkwalk -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4700 iso
 
-snmptable -Cib -v1 -c public -l noAuthNoPriv -u MD5DES -n "" localhost:4700 snmptargetaddrtable
-snmptable -Cib -v2c -c public -l noAuthNoPriv -u MD5DES -n "" localhost:4700 snmptargetaddrtable
+snmptable -Cib -v1 -c public  localhost:4700 snmptargetaddrtable
+snmptable -Cib -v2c -c public localhost:4700 snmptargetaddrtable
 snmptable -Cib -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4700 snmptargetaddrtable
 
 snmpwalk -v3 -l AuthNoPriv -u MD5 -a SHA -A MD5UserAuthPassword -n "" localhost:4700 system || echo OK
@@ -36,9 +52,12 @@ snmpget -v3 -l noAuthNoPriv -u MD5DES -n "wrong" localhost:4700 snmpEnableAuthen
 snmpwalk -v3 -l AuthNoPriv -u SHA -a SHA -A WrongUserAuthPassword -n "" localhost:4700 || echo OK
 snmpwalk -v3 -l AuthNoPriv -u MD5 -n "" localhost:4700 || echo OK
 
-snmpwalk -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4700 SNMPv2-MIB::snmpOutTraps.0
+snmpOutTraps=`snmpget -v1 -c public -Onqv localhost:4700 snmpOutTraps.0`
+test ${snmpOutTraps} -ne 1 || exit 1
+echo "OK, authentication failure trap sent"
 
-_deps/snmp_pp-build/test_app
+# start snmp_pp test_app too
+_deps/snmp_pp-build/test_app 127.0.0.1 get
 
 # pkill agent
 kill -s TERM %%
@@ -46,15 +65,21 @@ wait
 
 ls -lrta config
 
-examples/static_table/src/agent 4711 &
+# start as bg job
+examples/static_table/src/agent 4700 &
 sleep 1
 cat snmpv3_boot_counter
 
-snmpget -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4711 snmpEngineBoots.0
-snmpset -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4711 snmpEngineBoots.0 = 1 || echo OK
-snmpset -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4711 sysContact.0 = clausklein
-snmpwalk -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4711 system
-snmpwalk -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4711 snmpEngine
+snmpEngineBoots=`snmpget -v2c -c public -Onqv localhost:4700 snmpEngineBoots.0`
+test ${snmpEngineBoots} -eq 2 || exit 1
+echo "OK, second boot"
+
+snmpget -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4700 snmpEngineBoots.0
+snmpset -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4700 snmpEngineBoots.0 = 1 || echo OK
+snmpset -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4700 sysContact.0 = clausklein
+snmpwalk -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4700 system
+snmpwalk -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4700 snmpEngine
+snmpwalk -v3 -l noAuthNoPriv -u MD5DES -n "" localhost:4700 SNMPv2-MIB::snmpOutTraps.0
 
 # pkill agent
 kill -s TERM %%
